extend pa_usec_t to 64 bit
git-svn-id: file:///home/lennart/svn/public/pulseaudio/trunk@192 fefdeb5f-60dc-0310-8127-8f9354f1896f

+++ polyp/sample.h
@@ -67,8 +67,8 @@ struct pa_sample_spec {
     uint8_t channels;              /**< Audio channels. (1 for mono, 2 for stereo, ...) */
 };
 
-/** Type for usec specifications */
-typedef uint32_t pa_usec_t;
+/** Type for usec specifications. May be either 32 or 64 bit, depending on the architecture */
+typedef uint64_t pa_usec_t;
 
 /** Return the amount of bytes playback of a second of audio with the speicified sample type takes */
 size_t pa_bytes_per_second(const struct pa_sample_spec *spec);
